collapse changes commit 1) user can choose topics on the tasks page when his tasks count is equal to 0. 2) after user chose his task topics, tasks will generate using deepseek api in solo-leveling-ai service 3) after 30 seconds user chose his topics, page will be reloaded (in future will be refactored using hazelcast)
