--- src/main/java/com/sleepkqq/sololeveling/Application.java
@@ -1,18 +1,0 @@
-package com.sleepkqq.sololeveling;
-
-import com.vaadin.flow.component.page.AppShellConfigurator;
-import com.vaadin.flow.component.page.Push;
-import com.vaadin.flow.theme.Theme;
-import com.vaadin.flow.theme.lumo.Lumo;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@Push
-@SpringBootApplication
-@Theme(value = "my-app", variant = Lumo.DARK)
-public class Application implements AppShellConfigurator {
-
-  public static void main(String[] args) {
-    SpringApplication.run(Application.class, args);
-  }
-}
