--- src/main/java/com/sleepkqq/sololeveling/localization/LocaleConverter.java
@@ -1,20 +1,0 @@
-package com.sleepkqq.sololeveling.localization;
-
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-
-import java.util.Locale;
-
-@Converter(autoApply = true)
-public class LocaleConverter implements AttributeConverter<Locale, String> {
-
-  @Override
-  public String convertToDatabaseColumn(Locale locale) {
-    return locale == null ? null : locale.toLanguageTag();
-  }
-
-  @Override
-  public Locale convertToEntityAttribute(String dbData) {
-    return dbData == null ? null : Locale.forLanguageTag(dbData);
-  }
-}
