--- src/main/java/com/sleepkqq/sololeveling/localization/LocalizationException.java
@@ -1,11 +1,0 @@
-package com.sleepkqq.sololeveling.localization;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public class LocalizationException extends RuntimeException {
-
-  private final LocalizationMessage localizationMessage;
-}
