--- src/main/java/com/sleepkqq/sololeveling/model/TgAuthData.java
@@ -1,9 +1,0 @@
-package com.sleepkqq.sololeveling.model;
-
-
-public record TgAuthData(
-    TgWebAppData tgWebAppData,
-    String initData
-) {
-
-}
