--- src/main/java/com/sleepkqq/sololeveling/model/UserRole.java
@@ -1,21 +1,0 @@
-package com.sleepkqq.sololeveling.model;
-
-import org.springframework.security.core.GrantedAuthority;
-
-public enum UserRole implements GrantedAuthority {
-  USER,
-  ADMIN;
-
-  @Override
-  public String getAuthority() {
-    return name();
-  }
-
-  public boolean isUser() {
-    return this == USER || isAdmin();
-  }
-
-  public boolean isAdmin() {
-    return this == ADMIN;
-  }
-}
