+++ src/main/java/com/sleepkqq/sololeveling/ui/Application.java
@@ -1,0 +1,18 @@
+package com.sleepkqq.sololeveling.ui;
+
+import com.vaadin.flow.component.page.AppShellConfigurator;
+import com.vaadin.flow.component.page.Push;
+import com.vaadin.flow.theme.Theme;
+import com.vaadin.flow.theme.lumo.Lumo;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@Push
+@SpringBootApplication
+@Theme(value = "my-app", variant = Lumo.DARK)
+public class Application implements AppShellConfigurator {
+
+  public static void main(String[] args) {
+    SpringApplication.run(Application.class, args);
+  }
+}
