+++ src/main/java/com/sleepkqq/sololeveling/ui/localization/LocalizationException.java
@@ -1,0 +1,11 @@
+package com.sleepkqq.sololeveling.ui.localization;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class LocalizationException extends RuntimeException {
+
+  private final LocalizationMessage localizationMessage;
+}
