+++ src/main/java/com/sleepkqq/sololeveling/ui/model/TgAuthData.java
@@ -1,0 +1,9 @@
+package com.sleepkqq.sololeveling.ui.model;
+
+
+public record TgAuthData(
+    TgWebAppData tgWebAppData,
+    String initData
+) {
+
+}
