+++ src/main/java/com/sleepkqq/sololeveling/ui/model/TgUserData.java
@@ -1,0 +1,20 @@
+package com.sleepkqq.sololeveling.ui.model;
+
+import com.fasterxml.jackson.databind.PropertyNamingStrategies;
+import com.fasterxml.jackson.databind.annotation.JsonNaming;
+
+@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
+public record TgUserData(
+    long id,
+    String username,
+    String firstName,
+    String lastName,
+    String photoUrl,
+    String languageCode,
+    Boolean addedToAttachmentMenu,
+    Boolean allowsWriteToPm,
+    Boolean isBot,
+    Boolean isPremium
+) {
+
+}
