--- src/test/java/com/sleepkqq/sololeveling/ApplicationTest.java
@@ -1,12 +1,0 @@
-package com.sleepkqq.sololeveling;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApplicationTest {
-
-  @Test
-  void contextLoads() {
-  }
-}
