+++ src/test/java/com/sleepkqq/sololeveling/ui/ApplicationTest.java
@@ -1,0 +1,12 @@
+package com.sleepkqq.sololeveling.ui;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApplicationTest {
+
+  @Test
+  void contextLoads() {
+  }
+}
